rename permission module and add text and atom encoders
