+++ lib/guardian/permissions/text_encoding.ex
@@ -1,4 +1,4 @@
-defmodule Guardian.Permissions.JsonEncoding do
+defmodule Guardian.Permissions.TextEncoding do
   @behaviour Guardian.Permissions.PermissionEncoding
   use Bitwise
 
@@ -6,7 +6,7 @@ defmodule Guardian.Permissions.JsonEncoding do
     perms = Map.get(perm_set, type)
 
     for {k, v} <- perms, band(value, v) == v, into: [] do
-      k |> to_string() |> String.to_atom()
+      k |> to_string()
     end
   end
 
@@ -16,12 +16,12 @@ defmodule Guardian.Permissions.JsonEncoding do
   end
 
   defp encode_value(value, _perm_set, acc) when is_atom(value),
-    do: [value | acc]
+    do: [Atom.to_string(value) | acc]
 
   defp encode_value(value, _perm_set, acc) when is_binary(value),
-    do: [String.to_atom(value) | acc]
+    do: [value | acc]
 
   def decode(value, _type, _perm_set) when is_list(value) do
-    value |> Enum.map(&to_string/1) |> Enum.map(&String.to_atom/1)
+    value |> Enum.map(&String.to_atom/1)
   end
 end
